implement the order shop page
