+++ src/app/page.tsx
@@ -1,12 +1,10 @@
 import LandingPageComponent from "@/components/landing-page/LandingComponent";
-import RecentOrders from "@/components/dashboard/recent-order/RecentOrder";
+import RecentOrders from "@/components/dashboard/components/RecentOrder";
 
 export default function LadingPage() {
   return (
     <>
       <LandingPageComponent />
-      <RecentOrders/>
-
     </>
   );
 }
